[#60] 대본 생성 주제/소주제 입력 화면 디자인 수정
