+++ app/src/main/java/com/project/balpyo/Script/ScriptTopicFragment.kt
@@ -1,0 +1,49 @@
+package com.project.balpyo.Script
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.navigation.fragment.findNavController
+import com.project.balpyo.R
+import com.project.balpyo.Utils.MyApplication
+import com.project.balpyo.databinding.FragmentScriptTopicBinding
+
+class ScriptTopicFragment : Fragment() {
+
+    lateinit var binding: FragmentScriptTopicBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+
+        binding = FragmentScriptTopicBinding.inflate(layoutInflater)
+
+        binding.run {
+            buttonNext.setOnClickListener {
+                MyApplication.scriptTopic = editTextTopic.text.toString()
+                findNavController().navigate(R.id.scriptSubtopicFragment)
+            }
+        }
+
+        return binding.root
+    }
+
+    fun initToolBar() {
+        binding.run {
+            toolbar.buttonBack.visibility = View.VISIBLE
+            toolbar.buttonClose.visibility = View.INVISIBLE
+            toolbar.textViewTitle.text = "대본 생성"
+            toolbar.textViewPage.run {
+                visibility = View.VISIBLE
+                text = "2/5"
+            }
+            toolbar.buttonBack.setOnClickListener {
+                // 뒤로가기 버튼 클릭시 동작
+                findNavController().popBackStack()
+            }
+        }
+    }
+}
