+++ app/src/main/res/drawable/background_edittext.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <solid android:color="@color/white"/>
-    <stroke android:color="#F1F1F1"/>
+    <stroke android:color="@color/gray2"/>
     <stroke android:width="1dp"/>
 
-    <corners android:radius="8dp"/>
+    <corners android:radius="6dp"/>
 
 </shape>
