--- app/src/main/res/layout/fragment_script_subject.xml
@@ -1,157 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".Script.ScriptSubjectFragment">
-
-    <include
-        android:id="@+id/toolbar"
-        layout="@layout/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <LinearLayout
-        android:id="@+id/linearLayout_subject"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="20dp"
-        android:orientation="horizontal"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/toolbar">
-
-        <TextView
-            android:id="@+id/textView_subject"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:fontFamily="@font/pretendard_semi_bold"
-            android:text="발표 주제"
-            android:textColor="@color/text"
-            android:textSize="16sp" />
-
-        <ImageView
-            android:id="@+id/imageView_subject"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:src="@drawable/ic_essential" />
-    </LinearLayout>
-
-    <EditText
-        android:id="@+id/editText_subject"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="20dp"
-        android:background="@drawable/background_edittext"
-        android:ems="10"
-        android:fontFamily="@font/pretendard_regular"
-        android:hint="발표하려는 주제 (예: 발표 잘하는 법)"
-        android:inputType="text"
-        android:padding="10dp"
-        android:textColor="@color/text"
-        android:textColorHint="@color/gray3"
-        android:textSize="14sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/linearLayout_subject" />
-
-    <LinearLayout
-        android:id="@+id/linearLayout_subtopic"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="20dp"
-        android:orientation="horizontal"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/editText_subject">
-
-        <TextView
-            android:id="@+id/textView_subtopic"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:fontFamily="@font/pretendard_semi_bold"
-            android:text="발표 소주제"
-            android:textColor="@color/text"
-            android:textSize="16sp" />
-
-        <ImageView
-            android:id="@+id/imageView_subtopic"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:src="@drawable/ic_essential" />
-    </LinearLayout>
-
-    <EditText
-        android:id="@+id/editText_subtopic"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="20dp"
-        android:background="@drawable/background_edittext"
-        android:ems="10"
-        android:fontFamily="@font/pretendard_regular"
-        android:hint="발표의 소주제 (예: 긴장 풀기)"
-        android:inputType="text"
-        android:padding="10dp"
-        android:textColor="@color/text"
-        android:textColorHint="@color/gray3"
-        android:textSize="14sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/linearLayout_subtopic" />
-
-    <TextView
-        android:id="@+id/button_register"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="20dp"
-        android:fontFamily="@font/pretendard_semi_bold"
-        android:text="등록"
-        android:textColor="@color/primary"
-        app:layout_constraintBottom_toBottomOf="@+id/editText_subtopic"
-        app:layout_constraintEnd_toEndOf="@+id/editText_subtopic"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="@+id/editText_subtopic"
-        app:layout_constraintTop_toTopOf="@+id/editText_subtopic" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView_subtopic"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="10dp"
-        android:layout_marginEnd="20dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/editText_subtopic" />
-
-    <Button
-        android:id="@+id/button_next"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="50dp"
-        android:layout_marginEnd="20dp"
-        android:fontFamily="@font/pretendard_medium"
-        android:text="다음"
-        android:textSize="16sp"
-        app:cornerRadius="8dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/recyclerView_subtopic" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
